+++ src/main/java/terramine/common/item/armor/vanity/VanityArmor.java
@@ -8,7 +8,6 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
-import terramine.TerraMine;
 import terramine.common.item.armor.TerrariaArmor;
 
 import java.util.List;