+++ src/main/java/terramine/mixin/client/render/ItemInHandRendererMixin.java
@@ -18,7 +18,6 @@ import terramine.extensions.PlayerStorages;
 @Mixin(ItemInHandRenderer.class)
 public abstract class ItemInHandRendererMixin {
 
-    // todo: other players don't see on server, fix
     @WrapOperation(
             method = "renderHandsWithItems",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/ItemInHandRenderer;renderArmWithItem(Lnet/minecraft/client/player/AbstractClientPlayer;FFLnet/minecraft/world/InteractionHand;FLnet/minecraft/world/item/ItemStack;FLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V")