+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.1-1.20] - 2023-09-30
+### Changed
+- Magic Mirror sound now plays at the original and destination locations
+
+### Fixed
+- Players crashing on server when equipping an accessory with auto swing
+- Accessories and vanity not renderer for other players on server
+
 ## [1.7.0-1.20] - 2023-09-14
 ### Added
 - Meteorites (only spawn at exactly midnight, default 2% chance)